Replace symbol for Page Up, Page Down
Seems like the ⤓ glyph or its reverse is no longer covered by DejaVu
Sans Mono or proxied appropriately out of the box on Ubuntu 24.04, so
this commit replaces it with rotated versions of the symbols that are
still used for Home and End. This roughly matches the results obtained
before the rendering of the font changed.